Rename symbolic.py to terms.py
